+++ packages/pubsub/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.10"></a>
+## [1.0.10](https://github.com/aws/aws-amplify/compare/@aws-amplify/pubsub@1.0.9...@aws-amplify/pubsub@1.0.10) (2018-09-17)
+
+
+### Bug Fixes
+
+* **pubsub,api,appsync:** Use subscriber-id header (AppSync subscription) ([#1618](https://github.com/aws/aws-amplify/issues/1618)) ([03f2517](https://github.com/aws/aws-amplify/commit/03f2517))
+
+
+
+
+<a name="1.0.9"></a>
+## [1.0.9](https://github.com/aws/aws-amplify/compare/@aws-amplify/pubsub@1.0.8...@aws-amplify/pubsub@1.0.9) (2018-09-12)
+
+
+### Bug Fixes
+
+* **pubsub:** Better handling of multiple AppSync subscriptions ([#1611](https://github.com/aws/aws-amplify/issues/1611)) ([6c86500](https://github.com/aws/aws-amplify/commit/6c86500))
+
+
+
+
 <a name="1.0.8"></a>
 ## [1.0.8](https://github.com/aws/aws-amplify/compare/@aws-amplify/pubsub@1.0.7...@aws-amplify/pubsub@1.0.8) (2018-09-09)